requirements.txt
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
et-xmlfile==1.1.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
numpy==1.21.6
openpyxl==3.0.9
pandas==1.3.5
psycopg2==2.7.7
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
typing-extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
